+++ tests/test_date.py
@@ -6,7 +6,7 @@
 import unittest
 
 from ged4py.calendar import (
-    CalendarTypes, CalendarDate, FrenchDate, GregorianDate, HebrewDate, JulianDate,
+    CalendarType, CalendarDate, FrenchDate, GregorianDate, HebrewDate, JulianDate,
     CalendarDateVisitor
 )
 from ged4py.date import (
@@ -114,7 +114,7 @@ class TestDetailDate(unittest.TestCase):
         self.assertEqual(date.month, "OCT")
         self.assertEqual(date.month_num, 10)
         self.assertEqual(date.day, 9)
-        self.assertEqual(date.calendar, CalendarTypes.GREGORIAN)
+        self.assertEqual(date.calendar, CalendarType.GREGORIAN)
 
         date = GregorianDate(2017, "OCT", bc=True)
         self.assertEqual(date.year, 2017)
@@ -124,7 +124,7 @@ class TestDetailDate(unittest.TestCase):
         self.assertEqual(date.month, "OCT")
         self.assertEqual(date.month_num, 10)
         self.assertIsNone(date.day)
-        self.assertEqual(date.calendar, CalendarTypes.GREGORIAN)
+        self.assertEqual(date.calendar, CalendarType.GREGORIAN)
 
         date = GregorianDate(1699, "FEB", dual_year=1700)
         self.assertEqual(date.year, 1699)
@@ -134,7 +134,7 @@ class TestDetailDate(unittest.TestCase):
         self.assertEqual(date.month, "FEB")
         self.assertEqual(date.month_num, 2)
         self.assertIsNone(date.day)
-        self.assertEqual(date.calendar, CalendarTypes.GREGORIAN)
+        self.assertEqual(date.calendar, CalendarType.GREGORIAN)
 
         date = HebrewDate(5000)
         self.assertEqual(date.year, 5000)
@@ -143,7 +143,7 @@ class TestDetailDate(unittest.TestCase):
         self.assertIsNone(date.month)
         self.assertIsNone(date.month_num)
         self.assertIsNone(date.day)
-        self.assertEqual(date.calendar, CalendarTypes.HEBREW)
+        self.assertEqual(date.calendar, CalendarType.HEBREW)
 
         date = FrenchDate(1, "FRUC", 1)
         self.assertEqual(date.year, 1)
@@ -152,7 +152,7 @@ class TestDetailDate(unittest.TestCase):
         self.assertEqual(date.month, "FRUC")
         self.assertEqual(date.month_num, 12)
         self.assertEqual(date.day, 1)
-        self.assertEqual(date.calendar, CalendarTypes.FRENCH_R)
+        self.assertEqual(date.calendar, CalendarType.FRENCH_R)
 
         date = JulianDate(5, "JAN", bc=True)
         self.assertEqual(date.year, 5)
@@ -161,7 +161,7 @@ class TestDetailDate(unittest.TestCase):
         self.assertEqual(date.month, "JAN")
         self.assertEqual(date.month_num, 1)
         self.assertIsNone(date.day)
-        self.assertEqual(date.calendar, CalendarTypes.JULIAN)
+        self.assertEqual(date.calendar, CalendarType.JULIAN)
 
     def test_002_cal_date_key(self):
         """Test date.CalendarDate class."""
@@ -251,7 +251,7 @@ class TestDetailDate(unittest.TestCase):
         self.assertEqual(date.month_num, 5)
         self.assertEqual(date.day, 31)
         self.assertEqual(date.original, "31 MAY 2020")
-        self.assertEqual(date.calendar, CalendarTypes.GREGORIAN)
+        self.assertEqual(date.calendar, CalendarType.GREGORIAN)
 
         date = CalendarDate.parse("@#DGREGORIAN@ 10 MAR 1698/99")
         self.assertIsInstance(date, GregorianDate)
@@ -262,14 +262,14 @@ class TestDetailDate(unittest.TestCase):
         self.assertEqual(date.month_num, 3)
         self.assertEqual(date.day, 10)
         self.assertEqual(date.original, "@#DGREGORIAN@ 10 MAR 1698/99")
-        self.assertEqual(date.calendar, CalendarTypes.GREGORIAN)
+        self.assertEqual(date.calendar, CalendarType.GREGORIAN)
 
         date = CalendarDate.parse("10 MAR 1699/00")
         self.assertIsInstance(date, GregorianDate)
         self.assertEqual(date.year, 1699)
         self.assertEqual(date.dual_year, 1700)
         self.assertEqual(date.original, "10 MAR 1699/00")
-        self.assertEqual(date.calendar, CalendarTypes.GREGORIAN)
+        self.assertEqual(date.calendar, CalendarType.GREGORIAN)
 
         date = CalendarDate.parse("@#DJULIAN@ 100 B.C.")
         self.assertIsInstance(date, JulianDate)
@@ -279,7 +279,7 @@ class TestDetailDate(unittest.TestCase):
         self.assertIsNone(date.month_num)
         self.assertIsNone(date.day)
         self.assertEqual(date.original, "@#DJULIAN@ 100 B.C.")
-        self.assertEqual(date.calendar, CalendarTypes.JULIAN)
+        self.assertEqual(date.calendar, CalendarType.JULIAN)
 
         date = CalendarDate.parse("@#DFRENCH R@ 15 GERM 0001")
         self.assertIsInstance(date, FrenchDate)
@@ -289,7 +289,7 @@ class TestDetailDate(unittest.TestCase):
         self.assertEqual(date.month_num, 7)
         self.assertEqual(date.day, 15)
         self.assertEqual(date.original, "@#DFRENCH R@ 15 GERM 0001")
-        self.assertEqual(date.calendar, CalendarTypes.FRENCH_R)
+        self.assertEqual(date.calendar, CalendarType.FRENCH_R)
 
         date = CalendarDate.parse("@#DHEBREW@ 7 NSN 5000")
         self.assertIsInstance(date, HebrewDate)
@@ -299,7 +299,7 @@ class TestDetailDate(unittest.TestCase):
         self.assertEqual(date.month_num, 8)
         self.assertEqual(date.day, 7)
         self.assertEqual(date.original, "@#DHEBREW@ 7 NSN 5000")
-        self.assertEqual(date.calendar, CalendarTypes.HEBREW)
+        self.assertEqual(date.calendar, CalendarType.HEBREW)
 
         # cannot handle ROMAN
         with self.assertRaises(ValueError):